requirements.txt
contourpy==1.3.1
cycler==0.11.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
lightning-utilities==0.13.1
MarkupSafe==2.1.5
matplotlib==3.10.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
pip==25.0
pyparsing==3.2.0
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
setuptools==75.8.0
sip==6.10.0
six==1.16.0
sympy==1.13.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.6.2
torchvision==0.21.0+cu118
tornado==6.4.2
typing_extensions==4.12.2
unicodedata2==15.1.0
wheel==0.45.1
